requirements.txt
Django==4.2
django-filter==23.1
djangorestframework==3.14.0